break out set-in to it's own file
