+++ readme.md
@@ -16,7 +16,7 @@ npm install immutable-object-methods --save
 import {setIn} from 'immutable-object-methods';
 
 const input = {a: {b: 'c'}};
-const updated = setIn(input, ['a', 'd'], e);
+const updated = setIn(input, ['a', 'd'], 'e');
 
 console.log(input);
 console.log(updated);
